--- a/v2/go.mod
+++ b/v2/go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/google/go-cmp v0.4.0
 	github.com/google/uuid v1.1.1
 	github.com/hashicorp/golang-lru v0.5.3 // indirect
+	github.com/json-iterator/go v1.1.10
 	github.com/kr/text v0.2.0 // indirect
 	github.com/lightstep/tracecontext.go v0.0.0-20181129014701-1757c391b1ac
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect