Fix #609

This PR is based on the same idea of #614, but it's using [jsoniter library](https://pkg.go.dev/github.com/json-iterator/go#Iterator)

Coverage is pretty good (around 92%, we're just not hitting some error branches) and the benchmark results speak for themselves:

```
name                                                                               old time/op    new time/op    delta
Unmarshal/string_data_v0.3-8                                                         26.5µs ± 0%     9.2µs ± 2%  -65.23%  (p=0.008 n=5+5)
Unmarshal/nil_data_v0.3-8                                                            24.4µs ± 1%    10.0µs ± 0%  -58.78%  (p=0.016 n=5+4)
Unmarshal/string_data_v1.0-8                                                         28.4µs ± 0%    10.9µs ±14%  -61.79%  (p=0.008 n=5+5)
Unmarshal/base64_json_encoded_data_v1.0-8                                            15.9µs ± 0%     6.5µs ± 3%  -59.30%  (p=0.008 n=5+5)
Unmarshal/struct_data_v0.3-8                                                         27.2µs ± 1%     9.5µs ± 2%  -65.10%  (p=0.008 n=5+5)
Unmarshal/struct_data_v1.0-8                                                         29.0µs ± 0%    10.4µs ± 2%  -64.22%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8    29.0µs ± 0%    11.6µs ± 3%  -60.14%  (p=0.008 n=5+5)
Unmarshal/base64_xml_encoded_data_v1.0-8                                             16.7µs ± 1%     7.8µs ± 4%  -53.31%  (p=0.008 n=5+5)
Unmarshal/nil_data_v1.0-8                                                            26.5µs ± 0%    11.9µs ± 1%  -55.04%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8    27.4µs ± 1%     9.8µs ± 2%  -64.19%  (p=0.008 n=5+5)

name                                                                               old alloc/op   new alloc/op   delta
Unmarshal/string_data_v0.3-8                                                         9.62kB ± 0%    5.17kB ± 0%  -46.25%  (p=0.008 n=5+5)
Unmarshal/nil_data_v0.3-8                                                            8.09kB ± 0%    4.96kB ± 0%  -38.64%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-8                                                         9.62kB ± 0%    5.00kB ± 0%     ~     (p=0.079 n=4+5)
Unmarshal/base64_json_encoded_data_v1.0-8                                            5.37kB ± 0%    4.30kB ± 0%  -19.82%  (p=0.008 n=5+5)
Unmarshal/struct_data_v0.3-8                                                         9.67kB ± 0%    5.19kB ± 0%  -46.35%  (p=0.008 n=5+5)
Unmarshal/struct_data_v1.0-8                                                         9.67kB ± 0%    5.01kB ± 0%     ~     (p=0.079 n=4+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8    9.66kB ± 0%    6.46kB ± 0%  -33.20%  (p=0.008 n=5+5)
Unmarshal/base64_xml_encoded_data_v1.0-8                                             5.42kB ± 0%    4.32kB ± 0%  -20.23%  (p=0.008 n=5+5)
Unmarshal/nil_data_v1.0-8                                                            8.09kB ± 0%    4.96kB ± 0%  -38.64%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8    9.66kB ± 0%    6.45kB ± 0%  -33.21%  (p=0.008 n=5+5)

name                                                                               old allocs/op  new allocs/op  delta
Unmarshal/string_data_v0.3-8                                                            154 ± 0%        79 ± 0%  -48.70%  (p=0.008 n=5+5)
Unmarshal/nil_data_v0.3-8                                                               148 ± 0%        75 ± 0%  -49.32%  (p=0.008 n=5+5)
Unmarshal/string_data_v1.0-8                                                            154 ± 0%        77 ± 0%  -50.00%  (p=0.008 n=5+5)
Unmarshal/base64_json_encoded_data_v1.0-8                                               100 ± 0%        52 ± 0%  -48.00%  (p=0.008 n=5+5)
Unmarshal/struct_data_v0.3-8                                                            157 ± 0%        80 ± 0%  -49.04%  (p=0.008 n=5+5)
Unmarshal/struct_data_v1.0-8                                                            157 ± 0%        78 ± 0%  -50.32%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8       157 ± 0%        89 ± 0%  -43.31%  (p=0.008 n=5+5)
Unmarshal/base64_xml_encoded_data_v1.0-8                                                100 ± 0%        52 ± 0%  -48.00%  (p=0.008 n=5+5)
Unmarshal/nil_data_v1.0-8                                                               148 ± 0%        75 ± 0%  -49.32%  (p=0.008 n=5+5)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8       157 ± 0%        89 ± 0%  -43.31%  (p=0.008 n=5+5)
```

Plan for the followups:

* Use this in bindings and throughout various protocols to unmarshal structured events
* Try to implement the streaming writer using jsoniter `Stream`