@slinkydeveloper 2020-11-24
To justify the need for streaming (and the related complexity) i tried jsoniter on the old json code (just replacing all imports of `encoding/json` with `jsoniter`) and the results (old = master code, no_streaming = old code but with jsoniter replacing encoding/json, new = this PR):

```
name \ time/op                                                                     old.txt      no_streaming.txt  new.txt
Unmarshal/string_data_v0.3-8                                                       26.5µs ± 0%       16.3µs ± 3%   9.2µs ± 2%
Unmarshal/nil_data_v0.3-8                                                          24.4µs ± 1%       16.1µs ± 1%  10.0µs ± 0%
Unmarshal/string_data_v1.0-8                                                       28.4µs ± 0%       19.3µs ± 1%  10.9µs ±14%
Unmarshal/base64_json_encoded_data_v1.0-8                                          15.9µs ± 0%       11.1µs ± 3%   6.5µs ± 3%
Unmarshal/struct_data_v0.3-8                                                       27.2µs ± 1%       18.1µs ± 1%   9.5µs ± 2%
Unmarshal/struct_data_v1.0-8                                                       29.0µs ± 0%       19.4µs ± 1%  10.4µs ± 2%
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8  29.0µs ± 0%       19.5µs ± 2%  11.6µs ± 3%
Unmarshal/base64_xml_encoded_data_v1.0-8                                           16.7µs ± 1%       11.6µs ± 1%   7.8µs ± 4%
Unmarshal/nil_data_v1.0-8                                                          26.5µs ± 0%       17.9µs ± 1%  11.9µs ± 1%
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8  27.4µs ± 1%       17.3µs ± 2%   9.8µs ± 2%

name \ alloc/op                                                                    old.txt      no_streaming.txt  new.txt
Unmarshal/string_data_v0.3-8                                                       9.62kB ± 0%       7.81kB ± 0%  5.17kB ± 0%
Unmarshal/nil_data_v0.3-8                                                          8.09kB ± 0%       6.31kB ± 0%  4.96kB ± 0%
Unmarshal/string_data_v1.0-8                                                       9.62kB ± 0%       7.81kB ± 0%  5.00kB ± 0%
Unmarshal/base64_json_encoded_data_v1.0-8                                          5.37kB ± 0%       4.57kB ± 0%  4.30kB ± 0%
Unmarshal/struct_data_v0.3-8                                                       9.67kB ± 0%       7.82kB ± 0%  5.19kB ± 0%
Unmarshal/struct_data_v1.0-8                                                       9.67kB ± 0%       7.83kB ± 0%  5.01kB ± 0%
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8  9.66kB ± 0%       7.82kB ± 0%  6.46kB ± 0%
Unmarshal/base64_xml_encoded_data_v1.0-8                                           5.42kB ± 0%       4.75kB ± 0%  4.32kB ± 0%
Unmarshal/nil_data_v1.0-8                                                          8.09kB ± 0%       6.31kB ± 0%  4.96kB ± 0%
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8  9.66kB ± 0%       7.81kB ± 0%  6.45kB ± 0%

name \ allocs/op                                                                   old.txt      no_streaming.txt  new.txt
Unmarshal/string_data_v0.3-8                                                          154 ± 0%          134 ± 0%      79 ± 0%
Unmarshal/nil_data_v0.3-8                                                             148 ± 0%          129 ± 0%      75 ± 0%
Unmarshal/string_data_v1.0-8                                                          154 ± 0%          134 ± 0%      77 ± 0%
Unmarshal/base64_json_encoded_data_v1.0-8                                             100 ± 0%           92 ± 0%      52 ± 0%
Unmarshal/struct_data_v0.3-8                                                          157 ± 0%          135 ± 0%      80 ± 0%
Unmarshal/struct_data_v1.0-8                                                          157 ± 0%          135 ± 0%      78 ± 0%
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-8     157 ± 0%          135 ± 0%      89 ± 0%
Unmarshal/base64_xml_encoded_data_v1.0-8                                              100 ± 0%           94 ± 0%      52 ± 0%
Unmarshal/nil_data_v1.0-8                                                             148 ± 0%          129 ± 0%      75 ± 0%
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-8     157 ± 0%          135 ± 0%      89 ± 0%
```